requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
requests==2.26.0
Pillow==8.3.2
google-generativeai==0.3.1
PyPDF2==3.0.1



